+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -50,7 +50,7 @@ public let masterRuleList = RuleList(rules: [
     ForceTryRule.self,
     ForceUnwrappingRule.self,
     FunctionBodyLengthRule.self,
-    FunctionBodyWhitespaceCommentLinesRule.self,
+    FunctionBodyWhitespaceCommentRule.self,
     FunctionParameterCountRule.self,
     GenericTypeNameRule.self,
     IdentifierNameRule.self,