+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 
 #### Enhancements
 
-* Adds `function_body_whitespace_comment_lines` opt-in rule to prohibit empty and comment lines in function bodies.  
+* Adds `function_body_whitespace_comment` opt-in rule to prohibit empty and comment lines in function bodies.  
   [Mikhail Yakushin](https://github.com/driver733)
   [#2065](https://github.com/realm/SwiftLint/issues/2065)